+++ src/main/java/com/woowacourse/zzazanstagram/web/controller/comment/CommentController.java
@@ -1,17 +1,16 @@
-package com.woowacourse.zzazanstagram.web.controller.article;
+package com.woowacourse.zzazanstagram.web.controller.comment;
 
 import com.woowacourse.zzazanstagram.model.comment.domain.vo.CommentContents;
 import com.woowacourse.zzazanstagram.model.comment.dto.CommentResponse;
 import com.woowacourse.zzazanstagram.model.comment.service.CommentService;
 import com.woowacourse.zzazanstagram.model.member.MemberSession;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
-import java.util.List;
 
 @RestController
 public class CommentController {
@@ -28,18 +27,4 @@ public class CommentController {
         CommentResponse commentResponse = commentService.save(commentContents, articleId, memberSession.getEmail());
         return ResponseEntity.ok(commentResponse);
     }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<String> temp(MethodArgumentNotValidException e) {
-        BindingResult bindingResult = e.getBindingResult();
-        StringBuilder errorMessage = new StringBuilder();
-
-        List<ObjectError> allErrors = bindingResult.getAllErrors();
-        for (ObjectError allError : allErrors) {
-            errorMessage.append(allError.getDefaultMessage());
-        }
-
-        return ResponseEntity.badRequest().body(errorMessage.toString());
-    }
-
 }
